payload unit16

+++ src/boot/ksdk1.1.0/INA219.h
@@ -1,6 +1,6 @@
 void		initINA219(const uint8_t i2cAddress, uint16_t operatingVoltageMillivolts);
 WarpStatus	readSensorRegisterINA219(uint8_t deviceRegister, int numberOfBytes);
-WarpStatus	writeSensorRegisterINA219(uint8_t deviceRegister, uint8_t payloadBtye);
+WarpStatus	writeSensorRegisterINA219(uint8_t deviceRegister, uint16_t payloadBtye);
 WarpStatus 	configureSensorINA219(uint8_t payload_Config, uint8_t payload_Calibration);
 void		printSensorDataINA219(bool hexModeFlag);
 uint8_t		appendSensorDataINA219(uint8_t* buf);
